client: refactor how List inner logic
