+++ volumes_test.go
@@ -4,12 +4,6 @@ import (
 	"testing"
 )
 
-func TestVolumes(t *testing.T) {
-	var _ Taggable = (*Volume)(nil)
-	var _ syncCommand = (*ListVolumes)(nil)
-	var _ asyncCommand = (*ResizeVolume)(nil)
-}
-
 func TestVolume(t *testing.T) {
 	instance := &Volume{}
 	if instance.ResourceType() != "Volume" {
